requirements.txt
pyyaml>=6.0
rich>=13.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
nltk>=3.8
matplotlib>=3.7.0
jupyter>=1.0.0
ipython>=8.0.0